create seperate weather_spec.rb file to do the unit tests on weather and weather.rb file with the class Weather
